changes in cherry branch
There was mistake in the file ProductModal of main branch.
So created a new branch from that commit by typing git checkout -b <new_branch_name> <commit_hash> . this commit hash is of that commit where you want to go and fixed that mistake.
and after fixing  that mistake go and commit and push to newly created branch.
And to merge with main branch::::: coming to main branch and typed command git cherry-pick <commit_hash>
this commit_hash is of that commit which bug was fixed in newly created branch.
:: basically commit_hash is of new branch where bug was fixed.
Then if there was no conflict then it will be merged with main branch.
If there occurs conflict, resolve it and
you have to type command git cherry-pick --continue and then
 commit and push to main branch.